+++ test/browser-test-config/base-karma.js
@@ -9,6 +9,12 @@ module.exports = function (config) {
       'mocha'
     ],
 
+    mochaReporter: {
+      output: 'minimal'
+    },
+
+    reporters: ['mocha'],
+
     // list of files / patterns to load in the browser
     files: [
       'test/browser-test-config/browser-tests.test.js'